Adding navigation buttons component
